requirements.txt
fastapi
uvicorn[standard]
httpx
google-cloud-storage
